requirements.txt
pandas==0.24.0
numpy==1.19.0
scikit_learn==0.20.1
scipy==1.5.0
xgboost==0.80
kafka-python==1.4.3
